fix(gauntlet-runner,parallel-bounce): slop_score u32 → f64 in BounceJson
serde_json rejects float literals (e.g. 20.0) when deserializing into u32;
#[serde(default)] silently substituted 0, causing progress lines to always
display score=0 regardless of the actual bounce result.  Changing the field
to f64 matches the wire format and makes the terminal output truthful.

Root cause discovered during the data integrity cross-check: the ndjson log
and exported CSV were always correct (written by janitor bounce internally),
only the operator-visible progress display was lying.

+++ tools/gauntlet-runner/src/main.rs
@@ -100,8 +100,10 @@ struct AuthorLogin {
 /// is used only for the per-PR progress line emitted to stderr.
 #[derive(Debug, Deserialize, Default)]
 struct BounceJson {
+    /// Composite slop score.  Stored as `f64` in the bounce JSON; `u32` would
+    /// cause serde_json to reject the float literal and silently default to 0.
     #[serde(default)]
-    slop_score: u32,
+    slop_score: f64,
     #[serde(default)]
     unlinked_pr: u32,
     #[serde(default)]
